+++ assets/css/_generate-css.php
@@ -9,7 +9,7 @@
  * @uses        Custom CSS Styles Generator
  *
  * @since    1.0
- * @version  1.7
+ * @version  1.9.3
  */
 
 
@@ -113,7 +113,7 @@
 						 * For basic CSS file paths use this structure:
 						 * 'CUSTOM_CSS_FILE_SLUG'
 						 */
-						locate_template( $css_file_path, true );
+						require get_theme_file_path( $css_file_path );
 
 					}
 
@@ -137,7 +137,7 @@
  * Custom styles from skin editor
  */
 
-	locate_template( 'assets/css/_custom-styles.php', true ); //Must be in separate file for WordPress customizer
+	require get_theme_file_path( 'assets/css/_custom-styles.php' ); //Must be in separate file for WordPress customizer
 
 	$output .= "\r\n\r\n\r\n/**\r\n * Skin styles\r\n */\r\n\r\n" . wm_custom_styles() . "\r\n\r\n" . '/* End of file */';