+++ assets/css/_generate-rtl-css.php
@@ -10,7 +10,7 @@
  * @uses        require() instead of require_once() due to previous inclusion of files when building global CSS stylesheet
  *
  * @since    1.0
- * @version  1.7
+ * @version  1.9.3
  */
 
 
@@ -78,7 +78,7 @@
 						 * For basic CSS file paths use this structure:
 						 * 'CUSTOM_CSS_FILE_SLUG'
 						 */
-						locate_template( $css_file_path, true, false );
+						require get_theme_file_path( $css_file_path );
 
 					}
 
@@ -101,8 +101,6 @@
 /**
  * Custom styles from skin editor
  */
-
-	// locate_template( 'assets/css/_custom-styles.php', true ); //Must be in separate file for WordPress customizer
 
 	$output .= "\r\n\r\n\r\n/**\r\n * Skin styles\r\n */\r\n\r\n" . wm_custom_styles() . "\r\n\r\n" . '/* End of file */';